pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "h4kshot"
version = "0.1.0"
description = "Multiplatform screenshot/screen-recording utility with auto-upload to s.h4ks.com"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "h4ks-com"},
]
dependencies = [
    "requests>=2.28",
    "mss>=9.0",
    "Pillow>=10.0",
    "pystray>=0.19",
    "pynput>=1.7",
    "pyperclip>=1.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "responses>=0.23",
]

[project.scripts]
h4kshot = "h4kshot.app:main"

[project.gui-scripts]
h4kshot-gui = "h4kshot.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
